replace global macros with meaningful functions
Some functions (originally C macros in distorm.h) were implemented as
independent functions in the root of the Distorm class. Since they are not
tied explicitly to any structure, they could go unnoticed. Therefore, these
global functions were replaced with properties in the DInst structure. This
provides a more intuitive way to use the macros to adjust certain fields of
the DInst structure.